refactor: extract runtime state from AgentWorkspace
