+++ src/components/agent/use-agent-runtime-state-types.ts
@@ -1,0 +1,6 @@
+import type { AgentSessionContinuity } from "./agent-session-continuity";
+
+export type UseAgentRuntimeStateDependencies = {
+  continuity: AgentSessionContinuity | null;
+  selectedHermesSessionId: string | undefined;
+};
